pyproject.toml
[project]
name = "dialog-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "George Lu"}
]
readme = "README.md"
requires-python = "=3.10.12"
dependencies = [
    "boto3 (==1.38.14)",
    "openai (==0.27.7)",
    "pandas (==2.0.1)",
    "passlib (==1.7.4)",
    "protobuf (==4.22.5)",
    "pytest (==7.3.1)",
    "pytest-asyncio (==0.21.0)",
    "python-jose (==3.3.0)",
    "redis (==5.2.1)",
    "requests (==2.32.3)",
    "sqlalchemy (==2.0.13)",
    "twilio (==9.5.1)",
    "uvicorn (==0.34.3)",
    "fastapi (>=0.115.12,<0.116.0)",
    "langchain-community (>=0.3.24,<0.4.0)",
    "tensorflow (>=2.19.0,<3.0.0)",
    "mysql-connector-python (>=9.3.0,<10.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "numpy (==1.26.4)",
    "google-cloud (>=0.34.0,<0.35.0)",
    "google-cloud-translate (>=3.20.2,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
